1.1.4.6
Adds ECEF Conversions
Fixes NullRef bug occurring during move operations
