+++ CoordinateSharp/Coordinate.Assistant.cs
@@ -1,928 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Diagnostics;
-using System.Text.RegularExpressions;
 
 namespace CoordinateSharp
 {
-    /// <summary>
-    /// Coordinate formatting options for a Coordinate object.
-    /// </summary>
-    [Serializable]
-    public class CoordinateFormatOptions
-    {
-        /// <summary>
-        /// Set default values with the constructor.
-        /// </summary>
-        public CoordinateFormatOptions()
-        {
-            Format = CoordinateFormatType.Degree_Minutes_Seconds;
-            Round = 3;
-            Display_Leading_Zeros = false;
-            Display_Trailing_Zeros = false;
-            Display_Symbols = true;
-            Display_Degree_Symbol = true;
-            Display_Minute_Symbol = true;
-            Display_Seconds_Symbol = true;
-            Display_Hyphens = false;
-            Position_First = true;
-        }
-        /// <summary>
-        /// Coordinate format type.
-        /// </summary>
-        public CoordinateFormatType Format { get; set; }
-        /// <summary>
-        /// Rounds Coordinates to the set value.
-        /// </summary>
-        public int Round { get; set; }
-        /// <summary>
-        /// Displays leading zeros.
-        /// </summary>
-        public bool Display_Leading_Zeros { get; set; }
-        /// <summary>
-        /// Display trailing zeros.
-        /// </summary>
-        public bool Display_Trailing_Zeros { get; set; }
-        /// <summary>
-        /// Allow symbols to display.
-        /// </summary>
-        public bool Display_Symbols { get; set; }
-        /// <summary>
-        /// Display degree symbols.
-        /// </summary>
-        public bool Display_Degree_Symbol { get; set; }
-        /// <summary>
-        /// Display minute symbols.
-        /// </summary>
-        public bool Display_Minute_Symbol { get; set; }
-        /// <summary>
-        /// Display secons symbol.
-        /// </summary>
-        public bool Display_Seconds_Symbol { get; set; }
-        /// <summary>
-        /// Display hyphens between values.
-        /// </summary>
-        public bool Display_Hyphens { get; set; }
-        /// <summary>
-        /// Show coordinate position first.
-        /// Will show last if set 'false'.
-        /// </summary>
-        public bool Position_First { get; set; }
-    }
-    /// <summary>
-    /// Coordinate Format Types.
-    /// </summary>
-    [Serializable]
-    public enum CoordinateFormatType
-    {
-        /// <summary>
-        /// Decimal Degree Format
-        /// </summary>
-        /// <remarks>
-        /// Example: N 40.456 W 75.456
-        /// </remarks>
-        Decimal_Degree,
-        /// <summary>
-        /// Decimal Degree Minutes Format
-        /// </summary>
-        /// <remarks>
-        /// Example: N 40º 34.552' W 70º 45.408'
-        /// </remarks>
-        Degree_Decimal_Minutes,
-        /// <summary>
-        /// Decimal Degree Minutes Format
-        /// </summary>
-        /// <remarks>
-        /// Example: N 40º 34" 36.552' W 70º 45" 24.408'
-        /// </remarks>
-        Degree_Minutes_Seconds,
-        /// <summary>
-        /// Decimal Format
-        /// </summary>
-        /// <remarks>
-        /// Example: 40.57674 -70.46574
-        /// </remarks>
-        Decimal
-    }
-    internal class FormatFinder
-    {
-        //Add main to Coordinate and tunnel to Format class. Add private methods to format.
-        //WHEN PARSING NO EXCPETIONS FOR OUT OF RANGE ARGS WILL BE THROWN
-        public static bool TryParse(string coordString, out Coordinate c)
-        {
-            //Turn of eagerload for efficiency
-            EagerLoad eg = new EagerLoad();
-            eg.Cartesian = false;
-            eg.Celestial = false;
-            eg.UTM_MGRS = false;
+   
 
-            c = new Coordinate(eg);
-            string s = coordString;
-            s = s.Trim(); //Trim all spaces before and after string
-            double[] d;
-            //Try Signed Degree
-            if (TrySignedDegree(s, out d))
-            {
-                try
-                {
-                    c = new Coordinate(d[0], d[1], eg);
-                    return true;
-                }
-                catch
-                {//Parser failed try next method 
-                }
-            }
-            //Try Decimal Degree
-            if (TryDecimalDegree(s, out d))
-            {
-                try
-                {
-                    c = new Coordinate(d[0], d[1], eg);
-                    return true;
-                }
-                catch
-                {//Parser failed try next method 
-                }
-            }
-            //Try DDM
-            if (TryDegreeDecimalMinute(s, out d))
-            {
-                try
-                {
-                    //0 Lat Degree
-                    //1 Lat Minute
-                    //2 Lat Direction (0 = N, 1 = S)
-                    //3 Long Degree
-                    //4 Long Minute
-                    //5 Long Direction (0 = E, 1 = W)
-                    CoordinatesPosition latP = CoordinatesPosition.N;
-                    CoordinatesPosition lngP = CoordinatesPosition.E;
-                    if (d[2] != 0) { latP = CoordinatesPosition.S; }
-                    if (d[5] != 0) { lngP = CoordinatesPosition.W; }
-                    CoordinatePart lat = new CoordinatePart((int)d[0], d[1], latP, c);
-                    CoordinatePart lng = new CoordinatePart((int)d[3], d[4], lngP, c);
-                    c = new Coordinate(eg);
-                    c.Latitude = lat;
-                    c.Longitude = lng;
-                    return true;
-                }
-                catch
-                {//Parser failed try next method 
-                }
-            }
-            //Try DMS
-            if (TryDegreeMinuteSecond(s, out d))
-            {
-                try
-                {
-                    //0 Lat Degree
-                    //1 Lat Minute
-                    //2 Lat Second
-                    //3 Lat Direction (0 = N, 1 = S)
-                    //4 Long Degree
-                    //5 Long Minute
-                    //6 Long Second
-                    //7 Long Direction (0 = E, 1 = W)
-                    CoordinatesPosition latP = CoordinatesPosition.N;
-                    CoordinatesPosition lngP = CoordinatesPosition.E;
-                    if (d[3] != 0) { latP = CoordinatesPosition.S; }
-                    if (d[7] != 0) { lngP = CoordinatesPosition.W; }
-
-                    CoordinatePart lat = new CoordinatePart((int)d[0], (int)d[1], d[2], latP, c);
-                    CoordinatePart lng = new CoordinatePart((int)d[4], (int)d[5], d[6], lngP, c);
-                    c = new Coordinate(eg);
-                    c.Latitude = lat;
-                    c.Longitude = lng;
-                    return true;
-                }
-                catch
-                {//Parser failed try next method 
-                }
-            }
-
-            string[] um;
-            //Try MGRS
-            if (TryMGRS(s, out um))
-            {
-                try
-                {
-                    double zone = Convert.ToDouble(um[0]);
-                    double easting = Convert.ToDouble(um[3]);
-                    double northing = Convert.ToDouble(um[4]);
-                    MilitaryGridReferenceSystem mgrs = new MilitaryGridReferenceSystem(um[1], (int)zone, um[2], easting, northing);
-                    c = MilitaryGridReferenceSystem.MGRStoLatLong(mgrs);
-                    return true;
-                }
-                catch
-                {//Parser failed try next method 
-                }
-            }
-            //Try UTM
-            if (TryUTM(s, out um))
-            {
-                try
-                {
-                    double zone = Convert.ToDouble(um[0]);
-                    double easting = Convert.ToDouble(um[2]);
-                    double northing = Convert.ToDouble(um[3]);
-                    UniversalTransverseMercator utm = new UniversalTransverseMercator(um[1], (int)zone, easting, northing);
-                    c = UniversalTransverseMercator.ConvertUTMtoLatLong(utm);
-                    return true;
-                }
-                catch
-                {//Parser failed try next method 
-                }
-            }
-            //Try Cartesian
-            if (TryCartesian(s, out d))
-            {
-                try
-                {
-                    Cartesian cart = new Cartesian(d[0], d[1], d[2]);
-                    c = Cartesian.CartesianToLatLong(cart);
-                    return true;
-                }
-                catch
-                {//Parser failed try next method 
-                }
-            }
-            c = null;
-            return false;
-        }
-        private static bool TrySignedDegree(string s, out double[] d)
-        {
-            d = null;
-            if (Regex.Matches(s, @"[a-zA-Z]").Count != 0) { return false; } //Should contain no letters
-
-            string[] sA = SpecialSplit(s);
-            double lat;
-            double lng;
-
-            double degLat;
-            double minLat; //Minutes & MinSeconds
-            double secLat;
-
-            int signLat = 1;
-
-            double degLng;
-            double minLng; //Minutes & MinSeconds
-            double secLng;
-
-            int signLng = 1;
-
-            switch (sA.Count())
-            {
-                case 2:
-                    if (!double.TryParse(sA[0], out lat))
-                    { return false; }
-                    if (!double.TryParse(sA[1], out lng))
-                    { return false; }
-                    d = new double[] { lat, lng };
-                    return true;
-                case 4:
-                    if (!double.TryParse(sA[0], out degLat))
-                    { return false; }
-                    if (!double.TryParse(sA[1], out minLat))
-                    { return false; }
-                    if (!double.TryParse(sA[2], out degLng))
-                    { return false; }
-                    if (!double.TryParse(sA[3], out minLng))
-                    { return false; }
-
-                    if (degLat < 0) { signLat = -1; }
-                    if (degLng < 0) { signLng = -1; }
-                    if (minLat >= 60 || minLat < 0) { return false; } //Handle in parser as degree will be incorrect.
-                    if (minLng >= 60 || minLng < 0) { return false; } //Handle in parser as degree will be incorrect.
-                    lat = (Math.Abs(degLat) + (minLat / 60.0)) * signLat;
-                    lng = (Math.Abs(degLng) + (minLng / 60.0)) * signLng;
-                    d = new double[] { lat,lng };
-                    return true;
-                case 6:
-                    if (!double.TryParse(sA[0], out degLat))
-                    { return false; }
-                    if (!double.TryParse(sA[1], out minLat))
-                    { return false; }
-                    if (!double.TryParse(sA[2], out secLat))
-                    { return false; }
-                    if (!double.TryParse(sA[3], out degLng))
-                    { return false; }
-                    if (!double.TryParse(sA[4], out minLng))
-                    { return false; }
-                    if (!double.TryParse(sA[5], out secLng))
-                    { return false; }
-                    if (degLat < 0) { signLat = -1; }
-                    if (degLng < 0) { signLng = -1; }
-                    if (minLat >= 60 || minLat < 0) { return false; } //Handle in parser as degree will be incorrect.
-                    if (minLng >= 60 || minLng < 0) { return false; } //Handle in parser as degree will be incorrect.
-                    if (secLat >= 60 || secLat < 0) { return false; } //Handle in parser as degree will be incorrect.
-                    if (secLng >= 60 || secLng < 0) { return false; } //Handle in parser as degree will be incorrect.
-                    lat = (Math.Abs(degLat) + (minLat / 60.0) + (secLat/3600)) * signLat;
-                    lng = (Math.Abs(degLng) + (minLng / 60.0) + (secLng / 3600)) * signLng;
-                    d = new double[] { lat, lng };
-                    return true;
-                default:
-                    return false;
-            }         
-        }
-        private static bool TryDecimalDegree(string s, out double[] d)
-        {
-            d = null;
-            if (Regex.Matches(s, @"[a-zA-Z]").Count != 2) { return false; } //Should only contain 1 letter.
-
-            string[] sA = SpecialSplit(s);
-            if (sA.Count() == 2 || sA.Count() == 4)
-            {
-                double lat;
-                double lng;
-
-                double latR = 1; //Sets negative if South
-                double lngR = 1; //Sets negative if West
-
-                //Contact get brin directional indicator together with string
-                if (sA.Count() == 4)
-                {
-                    sA[0] += sA[1];
-                    sA[1] = sA[2] + sA[3];
-                }
-
-                //Find Directions
-                if (!sA[0].Contains("N") && !sA[0].Contains("n"))
-                {
-                    if (!sA[0].Contains("S") && !sA[0].Contains("s"))
-                    {
-                        return false;//No Direction Found
-                    }
-                    latR = -1;
-                }
-                if (!sA[1].Contains("E") && !sA[1].Contains("e"))
-                {
-                    if (!sA[1].Contains("W") && !sA[1].Contains("w"))
-                    {
-                        return false;//No Direction Found
-                    }
-                    lngR = -1;
-                }
-
-                sA[0] = Regex.Replace(sA[0], "[^0-9.]", "");
-                sA[1] = Regex.Replace(sA[1], "[^0-9.]", "");
-
-                if (!double.TryParse(sA[0], out lat))
-                { return false; }
-                if (!double.TryParse(sA[1], out lng))
-                { return false; }
-                lat *= latR;
-                lng *= lngR;
-                d = new double[] { lat, lng };
-                return true;
-            }
-
-            return false;
-        }
-        private static bool TryDegreeDecimalMinute(string s, out double[] d)
-        {
-            d = null;
-            if (Regex.Matches(s, @"[a-zA-Z]").Count != 2) { return false; } //Should only contain 1 letter.
-
-            string[] sA = SpecialSplit(s);
-            if (sA.Count() == 4 || sA.Count() == 6)
-            {
-                double latD;
-                double latMS;
-                double lngD;
-                double lngMS;
-
-                double latR = 0; //Sets 1 if South
-                double lngR = 0; //Sets 1 if West
-
-                //Contact get in order to combine directional indicator together with string
-                //Should reduce 6 items to 4
-                if (sA.Count() == 6)
-                {
-                    if (char.IsLetter(sA[0][0])) { sA[0] += sA[1]; sA[1] = sA[2]; }
-                    else if (char.IsLetter(sA[1][0])) { sA[0] += sA[1]; sA[1] = sA[2]; }
-                    else if (char.IsLetter(sA[2][0])) { sA[0] += sA[2]; }
-                    else { return false; }
-
-                    if (char.IsLetter(sA[3][0])) { sA[3] += sA[4]; sA[4] = sA[5]; }
-                    else if (char.IsLetter(sA[4][0])) { sA[3] += sA[4]; sA[4] = sA[5]; }
-                    else if (char.IsLetter(sA[5][0])) { sA[3] += sA[5]; }
-                    else { return false; }
-
-                    //Shift values for below logic
-                    sA[2] = sA[3];
-                    sA[3] = sA[4];
-                }
-
-                //Find Directions
-                if (!sA[0].Contains("N") && !sA[0].Contains("n") && !sA[1].Contains("N") && !sA[1].Contains("n"))
-                {
-                    if (!sA[0].Contains("S") && !sA[0].Contains("s") && !sA[1].Contains("S") && !sA[1].Contains("s"))
-                    {
-                        return false;//No Direction Found
-                    }
-                    latR = 1;
-                }
-                if (!sA[2].Contains("E") && !sA[2].Contains("e") && !sA[3].Contains("E") && !sA[3].Contains("e"))
-                {
-                    if (!sA[2].Contains("W") && !sA[2].Contains("w") && !sA[3].Contains("W") && !sA[3].Contains("w"))
-                    {
-                        return false;//No Direction Found
-                    }
-                    lngR = 1;
-                }
-
-                sA[0] = Regex.Replace(sA[0], "[^0-9.]", "");
-                sA[1] = Regex.Replace(sA[1], "[^0-9.]", "");
-                sA[2] = Regex.Replace(sA[2], "[^0-9.]", "");
-                sA[3] = Regex.Replace(sA[3], "[^0-9.]", "");
-
-                if (!double.TryParse(sA[0], out latD))
-                { return false; }
-                if (!double.TryParse(sA[1], out latMS))
-                { return false; }
-                if (!double.TryParse(sA[2], out lngD))
-                { return false; }
-                if (!double.TryParse(sA[3], out lngMS))
-                { return false; }
-
-                d = new double[] { latD, latMS, latR, lngD, lngMS, lngR };
-                return true;
-            }
-            return false;
-        }
-        private static bool TryDegreeMinuteSecond(string s, out double[] d)
-        {
-            d = null;
-            if (Regex.Matches(s, @"[a-zA-Z]").Count != 2) { return false; } //Should only contain 1 letter.
-
-            string[] sA = SpecialSplit(s);
-            if (sA.Count() == 6 || sA.Count() == 8)
-            {
-                double latD;
-                double latM;
-                double latS;
-                double lngD;
-                double lngM;
-                double lngS;
-
-                double latR = 0; //Sets 1 if South
-                double lngR = 0; //Sets 1 if West
-
-                //Contact get in order to combine directional indicator together with string
-                //Should reduce 8 items to 6
-                if (sA.Count() == 8)
-                {
-                    if (char.IsLetter(sA[0][0])) { sA[0] += sA[1]; sA[1] = sA[2]; sA[2] = sA[3]; }
-                    else if (char.IsLetter(sA[1][0])) { sA[0] += sA[1]; sA[1] = sA[2]; sA[2] = sA[3]; }
-                    else if (char.IsLetter(sA[3][0])) { sA[0] += sA[3]; }
-                    else { return false; }
-
-                    if (char.IsLetter(sA[4][0])) { sA[4] += sA[5]; sA[5] = sA[6]; sA[6] = sA[7]; }
-                    else if (char.IsLetter(sA[5][0])) { sA[4] += sA[5]; sA[5] = sA[6]; sA[6] = sA[7]; }
-                    else if (char.IsLetter(sA[7][0])) { sA[4] += sA[7]; }
-                    else { return false; }
-
-                    //Shift values for below logic
-                    sA[3] = sA[4];
-                    sA[4] = sA[5];
-                    sA[5] = sA[6];
-                }
-
-                //Find Directions
-                if (!sA[0].Contains("N") && !sA[0].Contains("n") && !sA[2].Contains("N") && !sA[2].Contains("n"))
-                {
-                    if (!sA[0].Contains("S") && !sA[0].Contains("s") && !sA[2].Contains("S") && !sA[2].Contains("s"))
-                    {
-                        return false;//No Direction Found
-                    }
-                    latR = 1;
-                }
-                if (!sA[3].Contains("E") && !sA[3].Contains("e") && !sA[5].Contains("E") && !sA[5].Contains("e"))
-                {
-                    if (!sA[3].Contains("W") && !sA[3].Contains("w") && !sA[5].Contains("W") && !sA[5].Contains("w"))
-                    {
-                        return false;//No Direction Found
-                    }
-                    lngR = 1;
-                }
-                sA[0] = Regex.Replace(sA[0], "[^0-9.]", "");
-                sA[1] = Regex.Replace(sA[1], "[^0-9.]", "");
-                sA[2] = Regex.Replace(sA[2], "[^0-9.]", "");
-                sA[3] = Regex.Replace(sA[3], "[^0-9.]", "");
-                sA[4] = Regex.Replace(sA[4], "[^0-9.]", "");
-                sA[5] = Regex.Replace(sA[5], "[^0-9.]", "");
-
-                if (!double.TryParse(sA[0], out latD))
-                { return false; }
-                if (!double.TryParse(sA[1], out latM))
-                { return false; }
-                if (!double.TryParse(sA[2], out latS))
-                { return false; }
-                if (!double.TryParse(sA[3], out lngD))
-                { return false; }
-                if (!double.TryParse(sA[4], out lngM))
-                { return false; }
-                if (!double.TryParse(sA[5], out lngS))
-                { return false; }
-
-                d = new double[] { latD, latM, latS, latR, lngD, lngM, lngS, lngR };
-                return true;
-            }
-            return false;
-        }
-        private static bool TryUTM(string s, out string[] utm)
-        {
-            utm = null;
-            string[] sA = SpecialSplit(s);
-            if (sA.Count() == 3 || sA.Count() == 4)
-            {
-                double zone;
-                string zoneL;
-                double easting;
-                double northing;
-
-                if (sA.Count() == 4)
-                {
-
-                    if (char.IsLetter(sA[0][0])) { sA[0] += sA[1]; sA[1] = sA[2]; sA[2] = sA[3]; }
-                    else if (char.IsLetter(sA[1][0])) { sA[0] += sA[1]; sA[1] = sA[2]; sA[2] = sA[3]; }
-                    else { return false; }
-                }
-                zoneL = new string(sA[0].Where(Char.IsLetter).ToArray());
-                if (zoneL == string.Empty) { return false; }
-                sA[0] = Regex.Replace(sA[0], "[^0-9.]", "");
-
-                if (!double.TryParse(sA[0], out zone))
-                { return false; }
-                if (!double.TryParse(sA[1], out easting))
-                { return false; }
-                if (!double.TryParse(sA[2], out northing))
-                { return false; }
-
-                utm = new string[] { zone.ToString(), zoneL, easting.ToString(), northing.ToString() };
-                return true;
-            }
-            return false;
-        }
-        private static bool TryMGRS(string s, out string[] mgrs)
-        {
-            mgrs = null;
-            string[] sA = SpecialSplit(s);
-            if (sA.Count() == 4 || sA.Count() == 5)
-            {
-                double zone;
-                string zoneL;
-                string diagraph;
-                double easting;
-                double northing;
-
-                if (sA.Count() == 5)
-                {
-                    if (char.IsLetter(sA[0][0])) { sA[0] += sA[1]; sA[1] = sA[2]; sA[2] = sA[3]; }
-                    else if (char.IsLetter(sA[1][0])) { sA[0] += sA[1]; sA[1] = sA[2]; sA[2] = sA[3]; }
-                    else { return false; }
-                }
-                zoneL = new string(sA[0].Where(Char.IsLetter).ToArray());
-                if (zoneL == string.Empty) { return false; }
-                sA[0] = Regex.Replace(sA[0], "[^0-9.]", "");
-                diagraph = sA[1];
-                if (!double.TryParse(sA[0], out zone))
-                { return false; }
-                if (!double.TryParse(sA[2], out easting))
-                { return false; }
-                if (!double.TryParse(sA[3], out northing))
-                { return false; }
-
-                mgrs = new string[] { zone.ToString(), zoneL, diagraph, easting.ToString(), northing.ToString() };
-                return true;
-            }
-            return false;
-        }
-        private static bool TryCartesian(string s, out double[] d)
-        {
-            d = null;
-            string[] sA = SpecialSplit(s);
-            if (sA.Count() == 3)
-            {
-                double x;
-                double y;
-                double z;
-                if (!double.TryParse(sA[0], out x))
-                { return false; }
-                if (!double.TryParse(sA[1], out y))
-                { return false; }
-                if (!double.TryParse(sA[2], out z))
-                { return false; }
-                d = new double[] { x, y, z };
-                return true;
-            }
-            return false;
-        }
-
-        private static string[] SpecialSplit(string s)
-        {
-            s = s.Replace("°", " ");
-            s = s.Replace("º", " ");
-            s = s.Replace("'", " ");
-            s = s.Replace("\"", " ");
-            s = s.Replace(",", " ");
-            s = s.Replace("mE", " ");
-            s = s.Replace("mN", " ");
-            return s.Split(new char[0], StringSplitOptions.RemoveEmptyEntries);
-        }
-    }
-    internal class FormatFinder_CoordPart
-    {
-        //Add main to Coordinate and tunnel to Format class. Add private methods to format.
-        //WHEN PARSING NO EXCPETIONS FOR OUT OF RANGE ARGS WILL BE THROWN
-        public static bool TryParse(string coordString, out CoordinatePart cp)
-        {
-            //Turn of eagerload for efficiency
-            EagerLoad eg = new EagerLoad();
-            int type = 0; //0 = unspecifed, 1 = lat, 2 = long;
-            eg.Cartesian = false;
-            eg.Celestial = false;
-            eg.UTM_MGRS = false;
-            cp = null;
-            Coordinate c = new Coordinate(eg);
-            string s = coordString;
-            s = s.Trim(); //Trim all spaces before and after string
-            double[] d;
-           
-            if(s[0] == ',')
-            {
-                type = 2;
-                s = s.Replace(",", "");
-                s = s.Trim();
-            }
-            if(s[0]=='*')
-            {
-                type = 1;
-                s = s.Replace("*", "");
-                s = s.Trim();
-            }
-     
-            if (TrySignedDegree(s, type, out d))
-            {
-                try
-                {
-                    switch (type)
-                    {
-                        case 0:
-                            //Attempt Lat first (default for signed)
-                            try
-                            {
-                                cp = new CoordinatePart(d[0], CoordinateType.Lat, c);
-                                return true;
-                            }
-                            catch
-                            {
-                                cp = new CoordinatePart(d[0], CoordinateType.Long, c);
-                                return true;
-                            }
-                        case 1:
-                            //Attempt Lat
-                            cp = new CoordinatePart(d[0], CoordinateType.Lat, c);
-                            return true;
-                        case 2:
-                            //Attempt long
-                            cp = new CoordinatePart(d[0], CoordinateType.Long, c);
-                            return true;
-                    }
-                }
-                catch
-                {
-                  //silent fail
-                }
-            }
-
-            //All other formats should contain 1 letter.
-            if (Regex.Matches(s, @"[a-zA-Z]").Count != 1) { return false; } //Should only contain 1 letter.
-            //Get Coord Direction
-            int direction = Find_Position(s);
-
-            if(direction == -1) { return false; //No direction found
-            }
-            //If Coordinate type int specified, look for mismatch
-            if(type == 1 && (direction == 1 || direction == 3)) { return false; //mismatch
-            }
-            if(type == 2 && (direction == 0 || direction == 2)) { return false; //mismatch
-            }
-            CoordinateType t;
-            if(direction == 0 || direction == 2) { t = CoordinateType.Lat; }
-            else { t = CoordinateType.Long; }
-
-            s = Regex.Replace(s, "[^0-9. ]", ""); //Remove directional character
-            s = s.Trim(); //Trim all spaces before and after string
-
-            //Try Decimal Degree with Direction
-            if (TryDecimalDegree(s, direction, out d))
-            {
-                try
-                {                   
-                    cp = new CoordinatePart(d[0],t, c);
-                    return true;
-                }
-                catch
-                {//Parser failed try next method 
-                }
-            }
-            //Try DDM
-            if (TryDegreeDecimalMinute(s, out d))
-            {
-                try
-                {
-                    //0  Degree
-                    //1  Minute
-                    //2  Direction (0 = N, 1 = E, 2 = S, 3 = W)                          
-                    cp = new CoordinatePart((int)d[0], d[1], (CoordinatesPosition)direction, c);                 
-                    return true;
-                }
-                catch
-                {
-                    //Parser failed try next method 
-                }
-            }
-            //Try DMS
-            if (TryDegreeMinuteSecond(s, out d))
-            {
-                try
-                {
-                    //0 Degree
-                    //1 Minute
-                    //2 Second
-                    //3 Direction (0 = N, 1 = E, 2 = S, 3 = W)                                     
-                    cp = new CoordinatePart((int)d[0], (int)d[1], d[2], (CoordinatesPosition)direction, c);
-                   
-                    return true;
-                }
-                catch
-                {//Parser failed try next method 
-                }
-            }
-            
-            return false;
-        }
-
-        private static bool TrySignedDegree(string s, int t, out double[] d)
-        {
-            d = null;
-            if (Regex.Matches(s, @"[a-zA-Z]").Count != 0) { return false; } //Should contain no letters
-
-            string[] sA = SpecialSplit(s);         
-            double deg;
-            double min; //Minutes & MinSeconds
-            double sec;
-
-            int sign = 1;
-            switch (sA.Count())
-            {
-                case 1:
-                    if (!double.TryParse(sA[0], out deg))
-                    { return false; }
-                    d = new double[] { deg };
-                    return true;
-                case 2:
-                    if (!double.TryParse(sA[0], out deg))
-                    { return false; }
-                    if (!double.TryParse(sA[1], out min))
-                    { return false; }
-               
-                    if(deg < 0) { sign = -1; }
-                    if(min >= 60 || min < 0) { return false; } //Handle in parser as degree will be incorrect.
-                    d = new double[] { (Math.Abs(deg) + (min / 60.0))*sign};
-                    return true;
-                case 3:
-                    if (!double.TryParse(sA[0], out deg))
-                    { return false; }
-                    if (!double.TryParse(sA[1], out min))
-                    { return false; }
-                    if (!double.TryParse(sA[2], out sec))
-                    { return false; }
-                    if (min >= 60 || min < 0) { return false; } //Handle in parser as degree will be incorrect.
-                    if (sec >= 60 || sec < 0) { return false; } //Handle in parser as degree will be incorrect.
-                   
-                    if (deg < 0) { sign = -1; }
-                    d = new double[] { (Math.Abs(deg) + (min / 60.0) + (sec / 3600.0))*sign };
-                    return true;
-                default:
-                    return false;
-            }         
-        }
-        private static bool TryDecimalDegree(string s, int direction, out double[] d)
-        {
-            d = null;
-            int sign = 1;
-            //S or W
-            if(direction == 2 || direction == 3)
-            {
-                sign = -1;
-            }           
-            double coord;
-
-            string[] sA = SpecialSplit(s);
-
-            if (sA.Count() == 1)
-            {
-                if (!double.TryParse(s, out coord))
-                { return false; }
-
-                coord *= sign;
-                d = new double[] { coord };
-                return true;
-            }
-
-            return false;
-        }
-        private static bool TryDegreeDecimalMinute(string s, out double[] d)
-        {
-            d = null;
-           
-            double deg;
-            double minSec;
-           
-            
-            string[] sA = SpecialSplit(s);
-            if (sA.Count() == 2)
-            {
-                if (!double.TryParse(sA[0], out deg))
-                { return false; }
-                if (!double.TryParse(sA[1], out minSec))
-                { return false; }
-
-                d = new double[] { deg,minSec};
-                return true;
-            }
-            return false;
-        }
-        private static bool TryDegreeMinuteSecond(string s, out double[] d)
-        {
-            d = null;
-          
-
-            double deg;
-            double min;
-            double sec;    
-
-            string[] sA = SpecialSplit(s);
-            if (sA.Count() == 3)
-            {
-
-                if (!double.TryParse(sA[0], out deg))
-                { return false; }
-                if (!double.TryParse(sA[1], out min))
-                { return false; }
-                if (!double.TryParse(sA[2], out sec))
-                { return false; }               
-
-                d = new double[] { deg, min, sec };
-                return true;
-            }
-            return false;
-        }
-
-        private static int Find_Position(string s)
-        {
-            //N=0
-            //E=1
-            //S=2
-            //W=3
-            //NOPOS = -1
-            //Find Directions
-
-            int part = -1;
-            if (s.Contains("N") || s.Contains("n"))
-            {
-                part = 0;
-            }
-            if (s.Contains("E") || s.Contains("e"))
-            {
-                part = 1;
-            }
-            if (s.Contains("S") || s.Contains("s"))
-            {
-                part = 2;
-               
-            }
-            if (s.Contains("W") || s.Contains("w"))
-            {
-                part = 3;
-            }
-            return part;
-        }
-        private static string[] SpecialSplit(string s)
-        {
-            s = s.Replace("°", " ");
-            s = s.Replace("º", " ");
-            s = s.Replace("'", " ");
-            s = s.Replace("\"", " ");
-            s = s.Replace(",", " ");
-            s = s.Replace("mE", " ");
-            s = s.Replace("mN", " ");
-            return s.Split(new char[0], StringSplitOptions.RemoveEmptyEntries);
-        }
-    }
+   
     /// <summary>
     /// Used for UTM/MGRS Conversions
     /// </summary>
@@ -1101,370 +185,42 @@ namespace CoordinateSharp
         /// </summary>
         W
     }
-    /// <summary>
-    /// Used to display a celestial condition for a specified date.
-    /// </summary>
-    [Serializable]
-    public enum CelestialStatus
-    {
-        /// <summary>
-        /// Celestial body rises and sets on the set day.
-        /// </summary>
-        RiseAndSet,
-        /// <summary>
-        /// Celestial body is down all day
-        /// </summary>
-        DownAllDay,
-        /// <summary>
-        /// Celestial body is up all day
-        /// </summary>
-        UpAllDay,
-        /// <summary>
-        /// Celestial body rises, but does not set on the set day
-        /// </summary>
-        NoRise,
-        /// <summary>
-        /// Celestial body sets, but does not rise on the set day
-        /// </summary>
-        NoSet
-    }
   
+ 
     /// <summary>
-    /// Turn on/off eager loading of certain properties.
-    /// </summary>
-    [Serializable]
-    public class EagerLoad
-    {
-        /// <summary>
-        /// Create an EagerLoad object
-        /// </summary>
-        public EagerLoad()
-        {      
-            Celestial = true;
-            UTM_MGRS = true;
-            Cartesian = true;
-        }
-
-        /// <summary>
-        /// Create an EagerLoad object with all options on or off
-        /// </summary>
-        /// <param name="isOn">Turns EagerLoad on or off</param>
-        public EagerLoad(bool isOn)
-        {
-            Celestial = isOn;
-            UTM_MGRS = isOn;
-            Cartesian = isOn;
-        }
-
-        /// <summary>
-        /// Create an EagerLoad object with only the specified flag options turned on.
-        /// </summary>
-        /// <param name="et">EagerLoadType</param>
-        public EagerLoad(EagerLoadType et)
-        {
-            Cartesian = false;
-            Celestial = false;
-            UTM_MGRS = false;
-
-            if (et.HasFlag(EagerLoadType.Cartesian))
-            {
-                Cartesian = true;
-            }
-            if(et.HasFlag(EagerLoadType.Celestial))
-            {
-                Celestial = true;
-            }
-            if(et.HasFlag(EagerLoadType.UTM_MGRS))
-            {
-                UTM_MGRS = true;
-            }
-        }
-
-        /// <summary>
-        /// Creates an EagerLoad object. Only the specified flags will be set to EagerLoad.
-        /// </summary>
-        /// <param name="et">EagerLoadType</param>
-        /// <returns>EagerLoad</returns>
-        public static EagerLoad Create(EagerLoadType et)
-        {
-            EagerLoad el = new EagerLoad(et);
-            return el;
-        }
-        
-        /// <summary>
-        /// Eager load celestial information.
-        /// </summary>
-        public bool Celestial { get; set; }      
-        /// <summary>
-        /// Eager load UTM and MGRS information
-        /// </summary>
-        public bool UTM_MGRS { get; set; }
-        /// <summary>
-        /// Eager load Cartesian information
-        /// </summary>
-        public bool Cartesian { get; set; }
-    }
-    /// <summary>
-    /// EagerLoad Enumerator
+    /// Coordinate type datum specification
     /// </summary>
     [Serializable]
     [Flags]
-    public enum EagerLoadType
+    public enum Coordinate_Datum
     {
+        /// <summary>
+        /// Lat Long GeoDetic
+        /// </summary>
+        LAT_LONG = 1,
         /// <summary>
         /// UTM and MGRS
         /// </summary>
-        UTM_MGRS=1,
+        UTM_MGRS = 2,
         /// <summary>
-        /// Celestial
+        /// ECEF
         /// </summary>
-        Celestial=2,
-        /// <summary>
-        /// Cartesian
-        /// </summary>
-        Cartesian=4
+        ECEF = 4,      
     }
-
+   
     /// <summary>
-    /// Contains distance values between two coordinates.
+    /// Cartesian Coordinate Type
     /// </summary>
-    [Serializable]
-    public class Distance 
-    {     
-        private double kilometers;        
-        private double miles;
-        private double feet;
-        private double meters;
-        private double bearing;
-        private double nauticalMiles;
-        
-        /// <summary>
-        /// Initializes a distance object using Haversine (Spherical Earth).
-        /// </summary>
-        /// <param name="c1">Coordinate 1</param>
-        /// <param name="c2">Coordinate 2</param>
-        public Distance(Coordinate c1, Coordinate c2)
-        {
-            Haversine(c1, c2);
-        }
-        /// <summary>
-        /// Initializes a distance object using Haversine (Spherical Earth) or Vincenty (Elliptical Earth).
-        /// </summary>
-        /// <param name="c1">Coordinate 1</param>
-        /// <param name="c2">Coordinate 2</param>
-        /// <param name="shape">Shape of earth</param>
-        public Distance(Coordinate c1, Coordinate c2, Shape shape)
-        {
-            if (shape == Shape.Sphere)
-            {
-                Haversine(c1, c2);
-            }
-            else
-            {
-                Vincenty(c1, c2);
-            }
-        }
-        /// <summary>
-        /// Initializes distance object based on distance in KM
-        /// </summary>
-        /// <param name="km">Kilometers</param>
-        public Distance(double km)
-        {
-            kilometers = km;
-            meters = km * 1000;
-            feet = meters * 3.28084;
-            miles = meters * 0.000621371;
-            nauticalMiles = meters * 0.0005399565;
-            bearing = 0;//None specified
-        }
-        /// <summary>
-        /// Initializaes distance object based on specified distance and measurement type
-        /// </summary>
-        /// <param name="distance">Distance</param>
-        /// <param name="type">Measurement type</param>
-        public Distance(double distance, DistanceType type)
-        {
-            bearing = 0;
-            switch(type)
-            {
-                case DistanceType.Feet:
-                    feet = distance;
-                    meters = feet * 0.3048;
-                    kilometers = meters / 1000;                  
-                    miles = meters * 0.000621371;
-                    nauticalMiles = meters * 0.0005399565;
-                    break;
-                case DistanceType.Kilometers:
-                    kilometers = distance;
-                    meters = kilometers * 1000;
-                    feet = meters * 3.28084;
-                    miles = meters * 0.000621371;
-                    nauticalMiles = meters * 0.0005399565;
-                    break;
-                case DistanceType.Meters:
-                    meters = distance;
-                    kilometers = meters / 1000;                 
-                    feet = meters * 3.28084;
-                    miles = meters * 0.000621371;
-                    nauticalMiles = meters * 0.0005399565;
-                    break;
-                case DistanceType.Miles:
-                    miles = distance;
-                    meters = miles * 1609.344;
-                    feet = meters * 3.28084;                 
-                    kilometers = meters / 1000;
-                    nauticalMiles = meters * 0.0005399565;                 
-                    break;
-                case DistanceType.NauticalMiles:
-                    nauticalMiles = distance;
-                    meters = nauticalMiles * 1852.001;
-                    feet = meters * 3.28084;
-                    kilometers = meters / 1000;
-                    miles = meters * 0.000621371;
-                    break;
-                default:
-                    kilometers = distance;
-                    meters = distance * 1000;
-                    feet = meters * 3.28084;
-                    miles = meters * 0.000621371;
-                    nauticalMiles = meters * 0.0005399565;                
-                    break;
-            }
-        }
-        private void Vincenty(Coordinate c1, Coordinate c2)
-        {
-            double lat1, lat2, lon1, lon2;
-            double d, crs12, crs21;
-      
-            lat1 = c1.Latitude.ToRadians();
-            lat2 = c2.Latitude.ToRadians();
-            lon1 = c1.Longitude.ToRadians();
-            lon2 = c2.Longitude.ToRadians();
-            //Ensure datums match between coords
-            if((c1.UTM.Equatorial_Radius != c2.UTM.Equatorial_Radius) || (c1.UTM.Inverse_Flattening!= c2.UTM.Inverse_Flattening))
-            {
-                throw new InvalidOperationException("The datum set does not match between Coordinate objects.");
-            }
-            double[] ellipse = new double[] { c1.UTM.Equatorial_Radius, c1.UTM.Inverse_Flattening };
-
-
-            // elliptic code
-            double[] cde = Distance_Assistant.Dist_Ell(lat1, -lon1, lat2, -lon2, ellipse);  // ellipse uses East negative
-            crs12 = cde[1] * (180 / Math.PI); //Bearing
-            crs21 = cde[2] * (180 / Math.PI); //Reverse Bearing
-            d = cde[0]; //Distance
-
-            bearing = crs21;
-            //reverseBearing = crs12;
-            meters = d;
-            kilometers = d/ 1000;         
-            feet = d * 3.28084;
-            miles = d * 0.000621371;
-            nauticalMiles = d * 0.0005399565;
-
-        }
-        
-        private void Haversine(Coordinate c1, Coordinate c2)
-        {
-            ////RADIANS
-            double nLat = c1.Latitude.ToDouble() * Math.PI / 180;
-            double nLong = c1.Longitude.ToDouble() * Math.PI / 180;
-            double cLat = c2.Latitude.ToDouble() * Math.PI / 180;
-            double cLong = c2.Longitude.ToDouble() * Math.PI / 180;
-
-            //Calcs
-            double R = 6371e3; //meters
-            double v1 = nLat;
-            double v2 = cLat;
-            double latRad = (c2.Latitude.ToDouble() - c1.Latitude.ToDouble()) * Math.PI / 180;
-            double longRad = (c2.Longitude.ToDouble() - c1.Longitude.ToDouble()) * Math.PI / 180;
-
-            double a = Math.Sin(latRad / 2.0) * Math.Sin(latRad / 2.0) +
-                Math.Cos(nLat) * Math.Cos(cLat) * Math.Sin(longRad / 2.0) * Math.Sin(longRad / 2.0);
-            double cl = 2 * Math.Atan2(Math.Sqrt(a), Math.Sqrt(1 - a));
-            double dist = R * cl;
-
-            //Get bearing
-            double y = Math.Sin(cLong - nLong) * Math.Cos(cLat);
-            double x = Math.Cos(nLat) * Math.Sin(cLat) -
-                    Math.Sin(nLat) * Math.Cos(cLat) * Math.Cos(cLong - nLong);
-            double brng = Math.Atan2(y, x) * (180/Math.PI); //Convert bearing back to degrees.
-            if(brng < 0) { brng -= 180; brng = Math.Abs(brng); }
-            kilometers = dist / 1000;
-            meters = dist;
-            feet = dist * 3.28084;
-            miles = dist * 0.000621371;
-            nauticalMiles = dist * 0.0005399565;
-            bearing = brng;
-        }
-        /// <summary>
-        /// Distance in Kilometers
-        /// </summary>
-        public double Kilometers
-        {
-            get { return kilometers; }
-        }
-        /// <summary>
-        /// Distance in Statute Miles
-        /// </summary>
-        public double Miles
-        {
-            get { return miles; }
-        }
-        /// <summary>
-        /// Distance in Nautical Miles
-        /// </summary>
-        public double NauticalMiles
-        {
-            get { return nauticalMiles; }
-        }
-        /// <summary>
-        /// Distance in Meters
-        /// </summary>
-        public double Meters
-        {
-            get { return meters; }
-        }
-        /// <summary>
-        /// Distance in Feet
-        /// </summary>
-        public double Feet
-        {
-            get { return feet; }
-        }
-        /// <summary>
-        /// Initial Bearing from Coordinate 1 to Coordinate 2
-        /// </summary>
-        public double Bearing
-        {
-            get { return bearing; }
-        }
-    }
-    /// <summary>
-    /// Distance measurement type
-    /// </summary>
-    public enum DistanceType
+    public enum CartesianType
     {
         /// <summary>
-        /// Distance in Meters
+        /// Spherical Cartesian
         /// </summary>
-        Meters,
+        Cartesian,
         /// <summary>
-        /// Distance in Kilometers
+        /// Earth Centered Earth Fixed
         /// </summary>
-        Kilometers,
-        /// <summary>
-        /// Distance in Feet
-        /// </summary>
-        Feet,
-        /// <summary>
-        /// Distance in Statute Miles
-        /// </summary>
-        Miles,
-        /// <summary>
-        /// Distance in Nautical Miles
-        /// </summary>
-        NauticalMiles
+        ECEF,
     }
     /// <summary>
     /// Used for easy read math functions
@@ -1495,6 +251,7 @@ namespace CoordinateSharp
     /// <summary>
     /// Earth Shape for Calculations.
     /// </summary>
+
     [Serializable]
     public enum Shape
     {
@@ -1507,194 +264,6 @@ namespace CoordinateSharp
         /// </summary>
         Ellipsoid
     }
-    [Serializable]
-    internal class Distance_Assistant
-    {
-        /// <summary>
-        /// Returns new geodetic coordinate in radians
-        /// </summary>
-        /// <param name="glat1">Latitude in Radians</param>
-        /// <param name="glon1">Longitude in Radians</param>
-        /// <param name="faz">Bearing</param>
-        /// <param name="s">Distance</param>
-        /// <param name="ellipse">Earth Ellipse Values</param>
-        /// <returns>double[]</returns>
-        public static double[] Direct_Ell(double glat1, double glon1, double faz, double s, double[] ellipse)
-        {
-            double EPS = 0.00000000005;//Used to determine if starting at pole.
-            double r, tu, sf, cf, b, cu, su, sa, c2a, x, c, d, y, sy = 0, cy = 0, cz = 0, e = 0;
-            double glat2, glon2, f;
-
-            //Determine if near pole
-            if ((Math.Abs(Math.Cos(glat1)) < EPS) && !(Math.Abs(Math.Sin(faz)) < EPS))
-            {
-                Debug.WriteLine("Warning: Location is at earth's pole. Only N-S courses are meaningful at this location.");
-            }
-
-
-            double a = ellipse[0];//Equitorial Radius
-            f = 1 / ellipse[1];//Flattening
-            r = 1 - f;
-            tu = r * Math.Tan(glat1);
-            sf = Math.Sin(faz);
-            cf = Math.Cos(faz);
-            if (cf == 0)
-            {
-                b = 0.0;
-            }
-            else
-            {
-                b = 2.0 * Math.Atan2(tu, cf);
-            }
-            cu = 1.0 / Math.Sqrt(1 + tu * tu);
-            su = tu * cu;
-            sa = cu * sf;
-            c2a = 1 - sa * sa;
-            x = 1.0 + Math.Sqrt(1.0 + c2a * (1.0 / (r * r) - 1.0));
-            x = (x - 2.0) / x;
-            c = 1.0 - x;
-            c = (x * x / 4.0 + 1.0) / c;
-            d = (0.375 * x * x - 1.0) * x;
-            tu = s / (r * a * c);
-            y = tu;
-            c = y + 1;
-            while (Math.Abs(y - c) > EPS)
-            {
-                sy = Math.Sin(y);
-                cy = Math.Cos(y);
-                cz = Math.Cos(b + y);
-                e = 2.0 * cz * cz - 1.0;
-                c = y;
-                x = e * cy;
-                y = e + e - 1.0;
-                y = (((sy * sy * 4.0 - 3.0) * y * cz * d / 6.0 + x) *
-                        d / 4.0 - cz) * sy * d + tu;
-            }
-
-            b = cu * cy * cf - su * sy;
-            c = r * Math.Sqrt(sa * sa + b * b);
-            d = su * cy + cu * sy * cf;
-
-            glat2 = ModM.ModLat(Math.Atan2(d, c));
-            c = cu * cy - su * sy * cf;
-            x = Math.Atan2(sy * sf, c);
-            c = ((-3.0 * c2a + 4.0) * f + 4.0) * c2a * f / 16.0;
-            d = ((e * cy * c + cz) * sy * c + y) * sa;
-            glon2 = ModM.ModLon(glon1 + x - (1.0 - c) * d * f);  //Adjust for IDL
-            //baz = ModM.ModCrs(Math.Atan2(sa, b) + Math.PI);
-            return new double[] { glat2, glon2 };
-        }
-        /// <summary>
-        /// Returns new geodetic coordinate in radians
-        /// </summary>
-        /// <param name="lat1">Latitude in radians</param>
-        /// <param name="lon1">Longitude in radians</param>
-        /// <param name="crs12">Bearing</param>
-        /// <param name="d12">Distance</param>
-        /// <returns>double[]</returns>
-        public static double[] Direct(double lat1, double lon1, double crs12, double d12)
-        {
-            var EPS = 0.00000000005;//Used to determine if near pole.
-            double dlon, lat, lon;
-            d12 = d12 * 0.0005399565; //convert meter to nm
-            d12 = d12 / (180 * 60 / Math.PI);//Convert to Radian
-            //Determine if near pole
-            if ((Math.Abs(Math.Cos(lat1)) < EPS) && !(Math.Abs(Math.Sin(crs12)) < EPS))
-            {
-                Debug.WriteLine("Warning: Location is at earth's pole. Only N-S courses are meaningful at this location.");
-            }
    
-            lat = Math.Asin(Math.Sin(lat1) * Math.Cos(d12) +
-                          Math.Cos(lat1) * Math.Sin(d12) * Math.Cos(crs12));
-            if (Math.Abs(Math.Cos(lat)) < EPS)
-            {
-                lon = 0.0; //endpoint a pole
-            }
-            else
-            {
-                dlon = Math.Atan2(Math.Sin(crs12) * Math.Sin(d12) * Math.Cos(lat1),
-                              Math.Cos(d12) - Math.Sin(lat1) * Math.Sin(lat));
-                lon = ModM.Mod(lon1 - dlon + Math.PI, 2 * Math.PI) - Math.PI;
-            }
-
-            return new double[] { lat, lon };
-        }
-        public static double[] Dist_Ell(double glat1, double glon1, double glat2, double glon2, double[] ellipse)
-        {
-            double a = ellipse[0]; //Equitorial Radius
-            double f = 1 / ellipse[1]; //Flattening
-
-            double r, tu1, tu2, cu1, su1, cu2, s1, b1, f1;
-            double x = 0, sx = 0, cx = 0, sy = 0, cy = 0, y = 0, sa = 0, c2a = 0, cz = 0, e = 0, c = 0, d = 0;
-            double EPS = 0.00000000005;
-            double faz, baz, s;
-            double iter = 1;
-            double MAXITER = 100;
-            if ((glat1 + glat2 == 0.0) && (Math.Abs(glon1 - glon2) == Math.PI))
-            {
-                Debug.WriteLine("Warning: Course and distance between antipodal points is undefined");
-                glat1 = glat1 + 0.00001; // allow algorithm to complete
-            }
-            if (glat1 == glat2 && (glon1 == glon2 || Math.Abs(Math.Abs(glon1 - glon2) - 2 * Math.PI) < EPS))
-            {
-                Debug.WriteLine("Warning: Points 1 and 2 are identical- course undefined");
-                //D
-                //crs12
-                //crs21
-                return new double[] { 0, 0, Math.PI };
-            }
-            r = 1 - f;
-            tu1 = r * Math.Tan(glat1);
-            tu2 = r * Math.Tan(glat2);
-            cu1 = 1.0 / Math.Sqrt(1.0 + tu1 * tu1);
-            su1 = cu1 * tu1;
-            cu2 = 1.0 / Math.Sqrt(1.0 + tu2 * tu2);
-            s1 = cu1 * cu2;
-            b1 = s1 * tu2;
-            f1 = b1 * tu1;
-            x = glon2 - glon1;
-            d = x + 1; // force one pass
-            while ((Math.Abs(d - x) > EPS) && (iter < MAXITER))
-            {
-                iter = iter + 1;
-                sx = Math.Sin(x);
-                cx = Math.Cos(x);
-                tu1 = cu2 * sx;
-                tu2 = b1 - su1 * cu2 * cx;
-                sy = Math.Sqrt(tu1 * tu1 + tu2 * tu2);
-                cy = s1 * cx + f1;
-                y = Math.Atan2(sy, cy);
-                sa = s1 * sx / sy;
-                c2a = 1 - sa * sa;
-                cz = f1 + f1;
-                if (c2a > 0.0)
-                {
-                    cz = cy - cz / c2a;
-                }
-                e = cz * cz * 2.0 - 1.0;
-                c = ((-3.0 * c2a + 4.0) * f + 4.0) * c2a * f / 16.0;
-                d = x;
-                x = ((e * cy * c + cz) * sy * c + y) * sa;
-                x = (1.0 - c) * x * f + glon2 - glon1;
-            }
-            faz = ModM.ModCrs(Math.Atan2(tu1, tu2));
-            baz = ModM.ModCrs(Math.Atan2(cu1 * sx, b1 * cx - su1 * cu2) + Math.PI);
-            x = Math.Sqrt((1 / (r * r) - 1) * c2a + 1);
-            x += 1;
-            x = (x - 2.0) / x;
-            c = 1.0 - x;
-            c = (x * x / 4.0 + 1.0) / c;
-            d = (0.375 * x * x - 1.0) * x;
-            x = e * cy;
-            s = ((((sy * sy * 4.0 - 3.0) * (1.0 - e - e) * cz * d / 6.0 - x) * d / 4.0 + cz) * sy * d + y) * c * a * r;
-
-            if (Math.Abs(iter - MAXITER) < EPS)
-            {
-                Debug.WriteLine("Warning: Distance algorithm did not converge");
-            }
-
-            return new double[] { s, faz, baz };
-        }
-    }
 
 }
